==> jupyter_notebooks/01_data_collection.ipynb <==
# **01 - Data Collection**

## Objectives

* Import Housing Price Data from Kaggle via API
* Save data: outputs/datasets/collection

## Inputs

* Kaggle JSON file - Authentication Token for the API

## Outputs

* Save House Price Dataset - outputs/datasets/collection/house_prices_records.csv



---

*Installing Requirements*

%pip install numpy==1.18.5 pandas==1.4.2 matplotlib==3.3.1 seaborn==0.11.0 pandas-profiling==3.1.0 plotly==4.12.0 ppscore==1.2.0 streamlit==0.85.0 feature-engine==1.0.2 imbalanced-learn==0.8.0 scikit-learn==0.23.0 xgboost==1.2.1 yellowbrick==1.3 Jinja2==3.1.1 MarkupSafe==2.0.1 protobuf==3.20 ipywidgets==8.0.2


# Change working directory

Change the working directory from its current folder to its parent folder

import os
current_dir = os.getcwd()
current_dir

Make the parent of the current directory the new current directory
* os.path.dirname() gets the parent directory
* os.chir() defines the new current directory

os.chdir(os.path.dirname(current_dir))
print("You set a new current directory")

Confirm the new current directory

current_dir = os.getcwd()
current_dir

# Fetch Data from Kaggle

Install the Kaggle package in order to fetch the data

%pip install kaggle

Set the Kakkle JSON Token for the session

import os
os.environ['KAGGLE_CONFIG_DIR'] = os.getcwd()
! chmod 600 kaggle.json

- Get the Kaggle Data
- Set the destination folder and download the data

KaggleDatasetPath = "codeinstitute/housing-prices-data"
DestinationFolder = "inputs/datasets/raw"   
! kaggle datasets download -d {KaggleDatasetPath} -p {DestinationFolder}

Unzip the data, delete the zip file and the kaggle token.

! unzip {DestinationFolder}/*.zip -d {DestinationFolder} \
  && rm {DestinationFolder}/*.zip \
  && rm kaggle.json

---

# Load and inspect the Housing Pricing Data

import pandas as pd
df = pd.read_csv(f"inputs/datasets/raw/house_prices_records.csv")
df.head()

Generate a Dataframe Summary

df.info()

# Convert Data Types as Required

Convert Area Values from Int to Float in order to have uniform data types.

df.info()

df[['1stFlrSF', 'BsmtFinSF1', 'BsmtUnfSF', 'GarageArea', 'GrLivArea', 'LotArea', 'OpenPorchSF', 'TotalBsmtSF', 'SalePrice', 'OverallCond', 'OverallQual']] = df[['1stFlrSF', 'BsmtFinSF1', 'BsmtUnfSF', 'GarageArea', 'GrLivArea', 'LotArea', 'OpenPorchSF', 'TotalBsmtSF', 'SalePrice', 'OverallCond', 'OverallQual']].astype(float)

df.info()

---

# Push files to Repo

* Create Outputs, Commit and Push to Repo

import os
try:

  os.makedirs(name='outputs/datasets/collection')
except Exception as e:
  print(e)

df.to_csv(f"outputs/datasets/collection/house_prices_records.csv",index=False)
